Split TextLayerBuilder into focused utility headers
